backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
google-generativeai==0.3.2
pydantic==2.5.3
python-dotenv==1.0.0
python-multipart==0.0.6
pypdf==3.17.4
sentence-transformers==2.7.0
faiss-cpu==1.8.0
numpy==1.26.3
